requirements.txt
asgiref     =        3.3.4
Django     =         3.2.1
djangorestframework= 3.12.4
pip               =  21.1.1
pytz              = 2021.1
setuptools        =  56.0.0
sqlparse          = 0.4.1
typing-extensions =  3.10.0.0
wheel             =  0.36.2
